Allow a prefix when specifying commits
This introduces the repo::refname syntax, in which case gitwebbot
will only look on repositories starting with the repo name. The following
are all acceptable:

  ::HEAD
  git::HEAD
  egit::3ea6fd234

The plain match-a-sha one will still look through all commits. Perhaps
it would be nice to make this default to a single repo, with a config
option.